requirements.txt
flask
python-dotenv
openai
pypdf2
requests
beautifulsoup4
lxml


